+++ FEXCore/Source/Interface/IR/IREmitter.cpp
@@ -9,7 +9,6 @@ $end_info$
 #include "Interface/IR/IREmitter.h"
 
 #include <FEXCore/IR/IR.h>
-#include <FEXCore/IR/IntrusiveIRList.h>
 #include <FEXCore/Utils/EnumUtils.h>
 #include <FEXCore/Utils/LogManager.h>
