+++ FEXCore/Source/Interface/IR/Passes/LongDivideRemovalPass.cpp
@@ -9,7 +9,6 @@ $end_info$
 #include "Interface/IR/IREmitter.h"
 #include "Interface/IR/PassManager.h"
 #include <FEXCore/IR/IR.h>
-#include <FEXCore/IR/IntrusiveIRList.h>
 #include <FEXCore/Utils/Profiler.h>
 
 #include <memory>
