+++ FEXCore/Source/Interface/IR/Passes/RedundantFlagCalculationElimination.cpp
@@ -10,7 +10,6 @@ $end_info$
 #include "Interface/IR/IREmitter.h"
 
 #include <FEXCore/IR/IR.h>
-#include <FEXCore/IR/IntrusiveIRList.h>
 #include <FEXCore/Utils/Profiler.h>
 
 #include "Interface/IR/PassManager.h"
