+++ FEXCore/Source/Interface/IR/Passes/ValueDominanceValidation.cpp
@@ -11,7 +11,6 @@ $end_info$
 #include "Interface/IR/PassManager.h"
 
 #include <FEXCore/IR/IR.h>
-#include <FEXCore/IR/IntrusiveIRList.h>
 #include <FEXCore/Utils/LogManager.h>
 #include <FEXCore/Utils/Profiler.h>
 #include <FEXCore/fextl/set.h>
