+++ FEXCore/Source/Interface/IR/RegisterAllocationData.h
