+++ FEXCore/include/FEXCore/Core/CoreState.h
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include "FEXCore/Core/X86Enums.h"
-#include "FEXCore/IR/IR.h"
+#include <FEXCore/Core/X86Enums.h>
+#include <FEXCore/IR/IR.h>
 #include <FEXCore/HLE/Linux/ThreadManagement.h>
 #include <FEXCore/Utils/CompilerDefs.h>
 #include <FEXCore/Utils/Telemetry.h>
